ImageServerPixelsID stored as a long instead of int
git-svn-id: file:///home/svn/shoola-omero/trunk@465 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/env/data/model/ImageSummary.java
@@ -62,7 +62,7 @@ public class ImageSummary
 	 * The OMEIS id's of this image's pixels.
 	 * The first element of the array always contains the default pixels.
 	 */
-	private int[]	imageServerPixelsID;
+	private long[]	imageServerPixelsID;
     
     /**
      * The default pixels of the image.
@@ -71,14 +71,14 @@ public class ImageSummary
 	
 	public ImageSummary() {}
 	
-	public ImageSummary(int id, String name, int[] imageServerPixelsID)
+	public ImageSummary(int id, String name, long[] imageServerPixelsID)
 	{
 		this.id = id;
 		this.name = name;
 		this.imageServerPixelsID = imageServerPixelsID;
 	}
     
-    public ImageSummary(int id, String name, int[] imageServerPixelsID,
+    public ImageSummary(int id, String name, long[] imageServerPixelsID,
                         PixelsDescription defaultPixels)
     {
         this.id  = id;
@@ -119,12 +119,12 @@ public class ImageSummary
 	}
 
 
-	public int[] getImageServerPixelsID()
+	public long[] getImageServerPixelsID()
 	{
 		return imageServerPixelsID;
 	}
 
-	public void setImageServerPixelsID(int[] imageServerPixelsID)
+	public void setImageServerPixelsID(long[] imageServerPixelsID)
 	{
 		this.imageServerPixelsID = imageServerPixelsID;
 	}
